refactor: search page SSR
